Implement queue insert and size functions
This commit implement q_insert_head q_insert_tail and q_size function.

The q_insert_head and q_insert_tail function will create a new node
and insert the node to the beginning or the end of the queue.

The q_size function will return the length of the queue.

Change-Id: I18615ac26f51660bb60a310328531c43c91b9223