Implement q_delete_mid function
To delete the mid of the queue, we can use two pointers to do it.
First, both pointers start at the first element in the queue.
The "slow" pointer moves one step at a time, and the "fast" pointer
moves two steps at a time.
Then, we walk through the queue with the condition
"fast != head && fast->next != head" .
After that, slow will point to the middle of the queue.

Change-Id: I7eeaa4ec99d48690beb3f897b816ed46f31ac3f3